* JarvisAndroidBranch: (59 commits)
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story - adding progress bar
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story - adding click event for hotel pane
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story - added check on build version to use appropriate fragmentmanager
  MOB-23497 - Jarvis Add Current Location to Location search results - added missing icon
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story - removing show maps activity
  MOB-23497 - Jarvis Add Current Location to Location search results
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story
  MOB-21846 : {New Hotel - Android} - Map - Hotels pinned on map - User Story
  MOB-23496 - Do not show voice icon when custom fields are available
  MOB-23398 : Android - Jarvis - The end user ends up on Hotel Search screen when going back from the itinerary screen straight after the booking.
  MOB-23281 - Jarvis - 'Display Message' rule is not displayed to the Android UI Reserve screen
  MOB-23407 : Android - Jarvis - Violation message is displayed even though it should not
  MOB-23407 : Android - Jarvis - Violation message is displayed even though it should not
  MOB-23407 : Android - Jarvis - Violation message is displayed even though it should not
  MOB-23434 - Able to deploy Cocnur Jarvis app alongside the play store app i.e Multibuild
  MOB-23408 - Jarvis - use spaces between the sentences for missing Required fields message
  MOB-22998 : Android - Jarvis - Parse error when booking with rates --> itinerary is not retrieved - added booking confirm pop-up cancel listener. - changed Logging setting from info to debug
  MOB-23384 - Jarvis - Minus sign is formed by 2 dashes for negative travel points
  MOB-23208 : Jarvis - Android - Handle custom text for 'Violation Justification' on the Violations screen
  ...

Conflicts:
	ConcurMobileAndroid.iml
	appcompat/project.properties
	apptentive-android/apptentive-android-sdk/.gitignore
	base/.gitignore
	base/build.gradle
	build.gradle
	core/.gitignore
	core/build.gradle
	core/src/com/concur/mobile/core/ConcurCore.java
	core/src/com/concur/mobile/core/travel/hotel/activity/RestHotelSearch.java
	core/src/com/concur/mobile/core/travel/hotel/activity/ShowHotelItinerary.java
	corporate/.gitignore
	corporate/AndroidManifest.xml
	corporate/build.gradle
	evature/.gitignore
	local.properties
	platform/.gitignore
	platform_ui_common/.gitignore
	platform_ui_common/res/values-v14/styles.xml
	platform_ui_common/src/com/concur/mobile/platform/ui/common/dialog/AlertDialogFragmentV1.java
	platform_ui_common/src/com/concur/mobile/platform/ui/common/util/ImageCache.java
	platform_ui_expense/.gitignore
	platform_ui_travel/.gitignore
	platform_ui_travel/build.gradle
	platform_ui_travel/res/layout/custom_dialog.xml
	platform_ui_travel/res/layout/hotel_booking.xml
	platform_ui_travel/res/values/strings.xml
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/activity/BaseActivity.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/hotel/activity/HotelBookingActivity.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/hotel/activity/HotelChoiceDetailsActivity.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/hotel/activity/HotelSearchAndResultActivity.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/hotel/fragment/CustomDialogFragment.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/loader/TravelCustomFieldsLoader.java
	platform_ui_travel/src/com/concur/mobile/platform/ui/travel/util/ViewUtil.java

+++ platform/src/com/concur/mobile/platform/travel/search/hotel/HotelRate.java
@@ -1,14 +1,13 @@
 package com.concur.mobile.platform.travel.search.hotel;
 
-import java.io.Serializable;
-
 import android.content.ContentResolver;
 import android.content.Context;
 import android.database.Cursor;
 import android.net.Uri;
-
 import com.concur.mobile.platform.travel.provider.Travel;
 import com.concur.mobile.platform.util.CursorUtil;
+
+import java.io.Serializable;
 
 /**
  * 
@@ -34,7 +33,7 @@ public class HotelRate implements Serializable {
     public String maxEnforcementLevelDesc;
     public URLInfo sellOptions;
     public int[] violationValueIds;
-    public Double travelPoints;
+    public Integer travelPoints;
     public boolean canRedeemTravelPointsAgainstViolations;
     /**
      * Contains the content Uri.
@@ -127,7 +126,7 @@ public class HotelRate implements Serializable {
         guaranteeSurcharge = CursorUtil.getStringValue(cursor, Travel.HotelRateDetailColumns.GUARANTEE_SURCHARGE);
         rateChangesOverStay = CursorUtil.getBooleanValue(cursor, Travel.HotelRateDetailColumns.RATE_CHANGES_OVERSTAY);
         maxEnforcementLevel = CursorUtil.getIntValue(cursor, Travel.HotelRateDetailColumns.MAX_ENF_LEVEL);
-        travelPoints = CursorUtil.getDoubleValue(cursor, Travel.HotelRateDetailColumns.TRAVEL_POINTS);
+        travelPoints = CursorUtil.getIntValue(cursor, Travel.HotelRateDetailColumns.TRAVEL_POINTS);
         canRedeemTravelPointsAgainstViolations = CursorUtil.getBooleanValue(cursor,
                 Travel.HotelRateDetailColumns.CAN_REDEEM_TP_AGAINST_VIOLATIONS);
